MIT-66: Fix bug in ADC stacking dimension.
In the dwi_pipeline the three ADC dimensions (adc_x, adc_y and adc_z)
are stacked as part of the processing. A bug was fixed where the
stackingwas being performed along the wrong axis.

Additionally, checks have been added to ensure that the ADC dimensions
are 3D and have the same shape.