MIT-73: Update to Nibabel=5.*
The updated major release of Nibabel introduced an API change that lead
to ValueError being raised when the nibabel1Image() was called with
int64 data (with no header or dtype arguments).

Some of the instances in mrimagetools where nibabel1Image() were called
were modified to include the dtype argument matching the data.
Specifically, instances where data comes from outside of the module,
tests that produce int64 data inputs, and tests where it is cumbersome
to identify that the dtype of the data is not int64.